refactor(apply): move the git and object-store roads into their fetchers
`_git_keep_last` had one caller (`GitSourceFetcher.fetch`) and
`acquire_object` had one (`ObjectStoreFetcher.fetch`), both reaching back
into the dispatcher that owned them. They move onto the fetchers as
`_keep_last` and `_acquire`, bodies unchanged apart from one fix: the
object road's per-entry cap is now the total lookup `FETCH_ENTRY_LIMITS[category]`
over the `FetchCategory` the front door already coerced, instead of a `.get`
defaulting to the file cap — the same way the git road's `file_limit=` line
already reads that table.

`FetchContext` and `scope_of` move to a new `apply/fetch_context.py`: both
fetchers and (next) the deliveries read them, so leaving them in the
dispatcher's module would close an import cycle.

The store-first/keep_last policy paragraphs move with the code they describe,
from the dispatcher's module docstring to `source_fetchers.py`.
